Add `cancel_limit_order` function
Clients need to pass in an `order_index` to cancel a limit order.
Because the Solana JSON-RPC `sendTransaction` call returns only a
transaction signature and not return data, I also needed to add a
`get_order_index` function that allows clients to get `order_index` with
a `ref_id`, which is a new field.